pyproject.toml
[project]
name = "ibm-rag-mvp"
version = "0.1.0"
description = "IBM Cloud RAG MVP with Streamlit, watsonx.ai, Milvus, and COS"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "IBM Lab MVP" }]

dependencies = [
  "streamlit>=1.37",
  "pydantic>=2.7",
  "pypdf>=4.2",
  "langchain-text-splitters>=0.2",
  "ibm-watsonx-ai>=1.1.8",
  "faiss-cpu>=1.7.4",
  "numpy>=1.26",
  "ibm-cos-sdk>=2.13",
  "tenacity>=8.2",
  "python-dotenv>=1.0",
  "requests>=2.31"
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


